pyproject.toml
[tool.poetry]
name = "django-amiqus"
version = "3.0"
description = "Django app for integration with Amiqus."
license = "MIT"
authors = ["YunoJuno <code@yunojuno.com>"]
maintainers = ["YunoJuno <code@yunojuno.com>"]
readme = "README.rst"
homepage = "https://github.com/yunojuno/django-amiqus"
repository = "https://github.com/yunojuno/django-amiqus"
documentation = "https://github.com/yunojuno/django-amiqus"
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 3.2",
    "Framework :: Django :: 4.0",
    "Framework :: Django :: 4.1",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
packages = [
    { include = "amiqus" }
]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.2 || ^4.0"
python-dateutil = "*"
requests =  "*"
simplejson = "*"

[tool.poetry.dev-dependencies]
black = "*"
coverage = "*"
flake8 = "*"
flake8-bandit = "*"
flake8-blind-except = "*"
flake8-docstrings = "*"
flake8-logging-format = "*"
flake8-print = "*"
freezegun = "*"
isort = "*"
mypy = "*"
pre-commit = "*"
pytest = "*"
pytest-cov = "*"
pytest-django = "*"
tox = "*"
types-requests = "*"
types-python-dateutil = "*"
types-simplejson = "*"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
